requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
numpy==2.3.3
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
